move AUTO files into new package, change "slep" method name to "sleep"
