+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autoModes/AUTO_SquareAndReverseSquare.java
@@ -1,10 +1,12 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.autoModes;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 
+import org.firstinspires.ftc.teamcode.Robot;
+
 @Autonomous
-public class AUTO_HexagonAndReverseHexagon extends OpMode {
+public class AUTO_SquareAndReverseSquare extends OpMode {
 
     Robot robot = new Robot(hardwareMap, gamepad1, gamepad2);
     public void init(){
@@ -18,41 +20,25 @@ public class AUTO_HexagonAndReverseHexagon extends OpMode {
 
     public void start(){
         robot.autoDrive.driveStraight(0.4, 24.0, 0.0);
-        slep(500);
+        sleep(500);
 
-        robot.autoDrive.turnToHeading(0.2, 60.0);
-        robot.autoDrive.holdHeading(0.2, 60.0, 0.5);
+        robot.autoDrive.turnToHeading(0.2, 90.0);
+        robot.autoDrive.holdHeading(0.2, 90.0, 0.5);
 
-        robot.autoDrive.driveStraight(0.4, 24.0, 60.0);
-        slep(500);
-
-        robot.autoDrive.turnToHeading(0.2, 120.0);
-        robot.autoDrive.holdHeading(0.2, 120.0, 0.5);
-
-        robot.autoDrive.driveStraight(0.4, 24.0, 120.0);
-        slep(500);
+        robot.autoDrive.driveStraight(0.4, 24.0, 90.0);
+        sleep(500);
 
         robot.autoDrive.turnToHeading(0.2, 180.0);
         robot.autoDrive.holdHeading(0.2, 180.0, 0.5);
 
         robot.autoDrive.driveStraight(0.4, 24.0, 180.0);
-        slep(500);
+        sleep(500);
 
-        robot.autoDrive.turnToHeading(0.2, -120.0);
-        robot.autoDrive.holdHeading(0.2, -120.0, 0.5);
+        robot.autoDrive.turnToHeading(0.2, -90.0);
+        robot.autoDrive.holdHeading(0.2, -90.0, 0.5);
 
-        robot.autoDrive.driveStraight(0.4, 24.0, -120.0);
-        slep(500);
-
-        robot.autoDrive.turnToHeading(0.2, -60.0);
-        robot.autoDrive.holdHeading(0.2, -60.0, 0.5);
-
-        robot.autoDrive.driveStraight(0.4, 24.0, -60.0);
-        slep(500);
-
-        robot.autoDrive.turnToHeading(0.2, 0.0);
-        robot.autoDrive.holdHeading(0.2, 0.0, 0.5);
-
+        robot.autoDrive.driveStraight(0.4, 24.0, -90.0);
+        sleep(500);
 
 
         //start doing all in reverse
@@ -60,25 +46,25 @@ public class AUTO_HexagonAndReverseHexagon extends OpMode {
         robot.autoDrive.holdHeading(0.2, 0.0, 0.5);
 
         robot.autoDrive.driveStraight(-0.4, 24.0, 0.0);
-        slep(500);
+        sleep(500);
 
         robot.autoDrive.turnToHeading(0.2, 90.0);
         robot.autoDrive.holdHeading(0.2, 90.0, 0.5);
 
         robot.autoDrive.driveStraight(-0.4, 24.0, 90.0);
-        slep(500);
+        sleep(500);
 
         robot.autoDrive.turnToHeading(0.2, 180.0);
         robot.autoDrive.holdHeading(0.2, 180.0, 0.5);
 
         robot.autoDrive.driveStraight(-0.4, 24.0, 180.0);
-        slep(500);
+        sleep(500);
 
         robot.autoDrive.turnToHeading(0.2, -90.0);
         robot.autoDrive.holdHeading(0.2, -90.0, 0.5);
 
         robot.autoDrive.driveStraight(-0.4, 24.0, -90.0);
-        slep(500);
+        sleep(500);
 
         robot.autoDrive.turnToHeading(0.2, 0.0);
         robot.autoDrive.holdHeading(0.2, 0.0, 0.5);
@@ -103,7 +89,7 @@ public class AUTO_HexagonAndReverseHexagon extends OpMode {
         robot.viperSlide.slideRot.setPower(0);
     }
 
-    public void slep(int ms){
+    public void sleep(int ms){
         android.os.SystemClock.sleep(ms);
     }
